organized generation a bit more
